+++ CHANGELOG.md
@@ -12,6 +12,9 @@
   did not already start with a capital letter, but this changes makes all type naming consistent.
   ([#2073](https://github.com/mozilla/uniffi-rs/issues/2073))
 
+- Python: Fix custom types generating invalid code when there are forward references.
+  ([#2067](https://github.com/mozilla/uniffi-rs/issues/2067))
+
 ### What's changed?
 - The async runtime can be specified for constructors/methods, this will override the runtime specified at the impl block level.